// src/app/api/swagger/models/api-quotation-session-application-dto-participant-info-dto.ts
/* tslint:disable */
/* eslint-disable */
import { ApiOrganizationCountryAltEnum } from '../models/api-organization-country';
export interface ApiQuotationSessionApplicationDtoParticipantInfoDto {
  Address?: string | null;
  Country?: ApiOrganizationCountryAltEnum;
  Email?: string | null;
  ForeignParticipantId?: string | null;
  Inn?: string | null;
  IsSmp?: boolean;
  Kpp?: string | null;
  Ogrn?: string | null;
  OrganizationGuid?: string | null;
  PhoneNumber?: string | null;
  PostAddress?: string | null;
}